requirements.txt
langchain
streamlit
google-generativeai
python-dotenv